organize editor package
