+++ packages/editor/src/jupyter/tooltip.js
